Extracted branches component, renamed Tracks to Application Paths
